🔧 FIX: Integer → BigInteger for telegram_id (PostgreSQL compatibility)

+++ models/website_user.py
@@ -1,7 +1,7 @@
 """
 Website User model for storing web registration data
 """
-from sqlalchemy import Column, Integer, String, DateTime, Boolean
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean
 from sqlalchemy.ext.declarative import declarative_base
 from datetime import datetime
 
@@ -11,7 +11,7 @@ class WebsiteUser(Base):
     __tablename__ = 'website_users'
     
     id = Column(Integer, primary_key=True, autoincrement=True)
-    telegram_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_id = Column(BigInteger, unique=True, nullable=False, index=True)
     email = Column(String(255), unique=True, nullable=False, index=True)
     password_hash = Column(String(255), nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
